requirements.txt
Flask==3.0.0
Whoosh==2.7.4
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
click==8.1.7
APScheduler==3.10.4
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
tabulate